Design

+++ src/components/taskList.tsx
@@ -2,6 +2,7 @@ import { useEffect, useState } from "react";
 import { useLoginUser } from "../context/accountManagementContext";
 import axios from "axios";
 import { useTasks } from "../context/tasksContext";
+import styles from "../styles/task.module.scss";
 
 const TaskList = () => {
   const [tasks, setTasks] = useTasks(); // 追加したタスク一覧を格納する
@@ -11,7 +12,7 @@ const TaskList = () => {
     // ユーザが登録したタスクを取得する関数
     const getTasks = async () => {
       let getTasksUrl: string =
-        "http://127.0.0.1:4000/users/" + user.id + "/tasks";
+        "https://task-tracker-ftp3.onrender.com/users/" + user.id + "/tasks";
       try {
         const response: any = await axios.get(getTasksUrl);
         setTasks(response.data);
@@ -24,10 +25,17 @@ const TaskList = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
+  // タスク名が変更されたらtasksを更新する関数
+  const tasksUpdate = async (newTaskName: string, index: number) => {
+    const updatedArray = [...tasks]; // オリジナル配列のコピーを作成
+    updatedArray[index].name = newTaskName; // 特定の要素を変更
+    setTasks(updatedArray);
+  };
+
   // 更新後のタスク一覧を取得する関数
   const getNewTasks = async () => {
     let getTasksUrl: string =
-      "http://127.0.0.1:4000/users/" + user.id + "/tasks";
+      "https://task-tracker-ftp3.onrender.com/users/" + user.id + "/tasks";
     try {
       const response: any = await axios.get(getTasksUrl);
       setTasks(response.data);
@@ -38,7 +46,7 @@ const TaskList = () => {
 
   // 登録済みのタスク名に変更があったら更新する関数
   const taskNameUpdate = async (newValue: string, task: any) => {
-    let updateUrl: string = "http://127.0.0.1:4000/tasks/" + task.id;
+    let updateUrl: string = "https://task-tracker-ftp3.onrender.com/tasks/" + task.id;
     try {
       await axios.patch(updateUrl, { name: newValue, user_id: user.id });
     } catch (error) {
@@ -57,7 +65,7 @@ const TaskList = () => {
 
     // 今日以降かどうか
     if (date > today) {
-      let updateUrl: string = "http://127.0.0.1:4000/tasks/" + task.id;
+      let updateUrl: string = "https://task-tracker-ftp3.onrender.com/tasks/" + task.id;
       try {
         await axios.patch(updateUrl, { deadline: date, user_id: user.id });
       } catch (error) {
@@ -71,7 +79,7 @@ const TaskList = () => {
 
   // 優先度に変更が合ったら更新する関数
   const taskPriorityUpdate = async (newValue: string, task: any) => {
-    let updateUrl: string = "http://127.0.0.1:4000/tasks/" + task.id;
+    let updateUrl: string = "https://task-tracker-ftp3.onrender.com/tasks/" + task.id;
     try {
       await axios.patch(updateUrl, {
         priority: Number(newValue),
@@ -85,7 +93,7 @@ const TaskList = () => {
 
   // 完了かどうかに変更が合ったら更新する関数
   const taskIsCompleteUpdate = async (newValue: string, task: any) => {
-    let updateUrl: string = "http://127.0.0.1:4000/tasks/" + task.id;
+    let updateUrl: string = "https://task-tracker-ftp3.onrender.com/tasks/" + task.id;
 
     if (newValue === "true") {
       try {
@@ -112,15 +120,19 @@ const TaskList = () => {
 
   // 課題を削除する関数
   const taskDelete = async (task: any) => {
-    let deleteUrl: string = "http://127.0.0.1:4000/tasks/" + task.id;
+    let result = confirm("本当に削除してよろしいですか？");
 
-    try {
-      await axios.delete(deleteUrl, { data: { user_id: user.id } });
-    } catch (error) {
-      alert("削除に失敗しました。");
+    if (result) {
+      let deleteUrl: string = "https://task-tracker-ftp3.onrender.com/tasks/" + task.id;
+
+      try {
+        await axios.delete(deleteUrl, { data: { user_id: user.id } });
+      } catch (error) {
+        alert("削除に失敗しました。");
+      }
+
+      getNewTasks();
     }
-
-    getNewTasks();
   };
 
   // 優先度(priority)でソートしたものを格納する
@@ -129,70 +141,106 @@ const TaskList = () => {
   );
 
   return (
-    <>
-      <div>
+    <div className={styles.listZone}>
+      <div className={styles.title}>
         <h1>タスク一覧</h1>
       </div>
       {/* タスクが存在しているかどうかによって表示内容を変更 */}
       {tasks.length > 0 ? (
-        <>
-          <div>
-            <h3>タイトル</h3>
-            <h3>締め切り</h3>
-            <h3>優先度</h3>
-            <h3>完了済み？</h3>
-          </div>
-          {sortedTasks.map((task: any, index: number) => {
-            return (
-              <div key={index}>
-                <input
-                  type="text"
-                  value={task.name}
-                  onChange={(e) => {
-                    taskNameUpdate(e.target.value, task);
-                  }}
-                />
-                <input
-                  type="date"
-                  value={task.deadline}
-                  onChange={(e) => {
-                    taskDeadlineUpdate(e.target.value, task);
-                  }}
-                />
-                <select
-                  value={task.priority}
-                  onChange={(e) => {
-                    taskPriorityUpdate(e.target.value, task);
-                  }}
-                >
-                  <option value={1}>高</option>
-                  <option value={2}>中</option>
-                  <option value={3}>低</option>
-                </select>
-                <select
-                  value={task.is_complete}
-                  onChange={(e) => {
-                    taskIsCompleteUpdate(e.target.value, task);
-                  }}
-                >
-                  <option value={"true"}>完了</option>
-                  <option value={"false"}>未完了</option>
-                </select>
-                <button
-                  onClick={() => {
-                    taskDelete(task);
-                  }}
-                >
-                  削除
-                </button>
-              </div>
-            );
-          })}
-        </>
+        <div>
+          <table>
+            <thead>
+              <tr>
+                <th>タイトル</th>
+                <th>締め切り</th>
+                <th>優先度</th>
+                <th>完了済み？</th>
+                <th></th>
+              </tr>
+            </thead>
+            <tbody>
+              {sortedTasks.map((task: any, index: number) => {
+                return (
+                  <tr key={index}>
+                    <td>
+                      <div>
+                        <input
+                          className={styles.taskNameZone}
+                          type="text"
+                          value={task.name}
+                          onChange={(e) => {
+                            tasksUpdate(e.target.value, index);
+                          }}
+                          onBlur={(e) => {
+                            taskNameUpdate(e.target.value, task);
+                          }}
+                        />
+                      </div>
+                    </td>
+                    <td>
+                      <div>
+                        <input
+                          type="date"
+                          value={task.deadline}
+                          onChange={(e) => {
+                            taskDeadlineUpdate(e.target.value, task);
+                          }}
+                        />
+                      </div>
+                    </td>
+                    <td>
+                      <div>
+                        <label className={styles.selectBox}>
+                          <select
+                            value={task.priority}
+                            onChange={(e) => {
+                              taskPriorityUpdate(e.target.value, task);
+                            }}
+                          >
+                            <option value={1}>高</option>
+                            <option value={2}>中</option>
+                            <option value={3}>低</option>
+                          </select>
+                        </label>
+                      </div>
+                    </td>
+                    <td>
+                      <div>
+                        <label className={styles.selectBox}>
+                          <select
+                            value={task.is_complete}
+                            onChange={(e) => {
+                              taskIsCompleteUpdate(e.target.value, task);
+                            }}
+                          >
+                            <option value={"true"}>完了</option>
+                            <option value={"false"}>未完了</option>
+                          </select>
+                        </label>
+                      </div>
+                    </td>
+                    <td>
+                      <div>
+                        <button
+                          className={styles.deleteButton}
+                          onClick={() => {
+                            taskDelete(task);
+                          }}
+                        >
+                          削除
+                        </button>
+                      </div>
+                    </td>
+                  </tr>
+                );
+              })}
+            </tbody>
+          </table>
+        </div>
       ) : (
         <></>
       )}
-    </>
+    </div>
   );
 };
